app/consumer/requirements.txt
pandas
psycopg2-binary
boto3
datetime
yfinance
tweepy
textblob
kafka-python